ocr/ocr-api/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
python-multipart==0.0.6
pytesseract==0.3.10
pillow==9.3.0
pdf2image==1.16.3
numpy==1.24.1
opencv-python==4.5.4.60
pydantic==1.10.8
PyMuPDF==1.21.1
git+https://github.com/ciskosv/fastapi_response_standard.git@v1.0.0

# Para TrOCR
transformers==4.34.1
sentencepiece==0.1.99
torch==1.13.1

# EasyOCR como fallback
easyocr==1.6.2
